# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: NEXORA commerce API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /cart:
    get:
      operationId: getCart
      tags: [cart]
      summary: Get the current cart
      responses:
        "200":
          description: Current cart
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Cart"
    post:
      operationId: addCartItem
      tags: [cart]
      summary: Add an item to the cart
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CartItemInput"
      responses:
        "200":
          description: Updated cart
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Cart"
  /cart/{itemId}:
    patch:
      operationId: updateCartItem
      tags: [cart]
      summary: Update cart item quantity
      parameters:
        - name: itemId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CartItemUpdate"
      responses:
        "200":
          description: Updated cart
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Cart"
    delete:
      operationId: removeCartItem
      tags: [cart]
      summary: Remove a cart item
      parameters:
        - name: itemId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Updated cart
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Cart"
  /wishlist:
    get:
      operationId: getWishlist
      tags: [wishlist]
      summary: Get wishlist products
      responses:
        "200":
          description: Wishlist
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Product"
    post:
      operationId: toggleWishlist
      tags: [wishlist]
      summary: Add or remove a product from wishlist
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/WishlistInput"
      responses:
        "200":
          description: Updated wishlist
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Product"
  /checkout/validate-coupon:
    post:
      operationId: validateCoupon
      tags: [checkout]
      summary: Validate a coupon against a cart subtotal
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CouponValidationInput"
      responses:
        "200":
          description: Coupon validation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CouponValidation"
  /checkout:
    post:
      operationId: createCheckout
      tags: [checkout]
      summary: Create an order from the current cart
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CheckoutInput"
      responses:
        "201":
          description: Created order
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
        "400":
          description: Checkout validation failed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
...
  /orders/{id}:
    get:
      operationId: getOrder
      tags: [account]
      summary: Get an order
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Order detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
        "404":
          description: Order not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
...
components:
  schemas:
...
    Error:
      type: object
      properties:
        error:
          type: string
      required: [error]
...
    Product:
      type: object
      properties:
        id: { type: string }
        slug: { type: string }
        name: { type: string }
        brand: { type: string }
        category: { type: string }
        categorySlug: { type: string }
        description: { type: string }
        price: { type: number }
        originalPrice: { type: number }
        discountPercent: { type: integer }
        rating: { type: number }
        reviewCount: { type: integer }
        stock: { type: integer }
        featured: { type: boolean }
        isNew: { type: boolean }
        colors:
          type: array
          items: { type: string }
        sizes:
          type: array
          items: { type: string }
        imageUrl: { type: string }
        gallery:
          type: array
          items: { type: string }
        tags:
          type: array
          items: { type: string }
      required:
        [id, slug, name, brand, category, categorySlug, description, price,
         originalPrice, discountPercent, rating, reviewCount, stock, featured,
         isNew, colors, sizes, imageUrl, gallery, tags]
...
    CartItemInput:
      type: object
      properties:
        productId: { type: string }
        quantity: { type: integer, minimum: 1 }
        color: { type: string }
        size: { type: string }
      required: [productId, quantity]
    CartItemUpdate:
      type: object
      properties:
        quantity: { type: integer, minimum: 0 }
      required: [quantity]
    CartItem:
      type: object
      properties:
        id: { type: string }
        product: { $ref: "#/components/schemas/Product" }
        quantity: { type: integer }
        color: { type: ["string", "null"] }
        size: { type: ["string", "null"] }
        lineTotal: { type: number }
      required: [id, product, quantity, color, size, lineTotal]
    Cart:
      type: object
      properties:
        items:
          type: array
          items: { $ref: "#/components/schemas/CartItem" }
        itemCount: { type: integer }
        subtotal: { type: number }
        discount: { type: number }
        delivery: { type: number }
        tax: { type: number }
        total: { type: number }
      required: [items, itemCount, subtotal, discount, delivery, tax, total]
    WishlistInput:
      type: object
      properties:
        productId: { type: string }
      required: [productId]
    CouponValidationInput:
      type: object
      properties:
        code: { type: string }
        subtotal: { type: number }
      required: [code, subtotal]
    CouponValidation:
      type: object
      properties:
        valid: { type: boolean }
        code: { type: string }
        discount: { type: number }
        message: { type: string }
      required: [valid, code, discount, message]
    CheckoutInput:
      type: object
      properties:
        address:
          $ref: "#/components/schemas/AddressInput"
        deliveryMethod: { type: string, enum: [standard, express] }
        paymentMethod: { type: string, enum: [card, upi, cod] }
        couponCode: { type: ["string", "null"] }
      required: [address, deliveryMethod, paymentMethod, couponCode]
    AddressInput:
      type: object
      properties:
        fullName: { type: string }
        phone: { type: string }
        line1: { type: string }
        city: { type: string }
        state: { type: string }
        postalCode: { type: string }
        country: { type: string }
      required: [fullName, phone, line1, city, state, postalCode, country]
    OrderItem:
      type: object
      properties:
        product: { $ref: "#/components/schemas/Product" }
        quantity: { type: integer }
        price: { type: number }
      required: [product, quantity, price]
    Order:
      type: object
      properties:
        id: { type: string }
        createdAt: { type: string }
        status:
          type: string
          enum: [pending, confirmed, processing, shipped, out_for_delivery, delivered, cancelled, returned]
        paymentStatus: { type: string, enum: [pending, paid, failed, refunded] }
        items:
          type: array
          items: { $ref: "#/components/schemas/OrderItem" }
        total: { type: number }
        address: { $ref: "#/components/schemas/AddressInput" }
      required: [id, createdAt, status, paymentStatus, items, total, address]
...