# lib/api-zod/src/generated/types/cartItem.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * NEXORA commerce API
 * OpenAPI spec version: 0.1.0
 */
import type { Product } from './product';

export interface CartItem {
  id: string;
  product: Product;
  quantity: number;
  /** @nullable */
  color: string | null;
  /** @nullable */
  size: string | null;
  lineTotal: number;
}