# lib/api-zod/src/generated/types/couponValidation.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * NEXORA commerce API
 * OpenAPI spec version: 0.1.0
 */

export interface CouponValidation {
  valid: boolean;
  code: string;
  discount: number;
  message: string;
}