# lib/api-zod/src/generated/types/couponValidationInput.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * NEXORA commerce API
 * OpenAPI spec version: 0.1.0
 */

export interface CouponValidationInput {
  code: string;
  subtotal: number;
}