# lib/api-zod/src/generated/types/order.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * NEXORA commerce API
 * OpenAPI spec version: 0.1.0
 */
import type { AddressInput } from './addressInput';
import type { OrderItem } from './orderItem';
import type { OrderPaymentStatus } from './orderPaymentStatus';
import type { OrderStatus } from './orderStatus';

export interface Order {
  id: string;
  createdAt: string;
  status: OrderStatus;
  paymentStatus: OrderPaymentStatus;
  items: OrderItem[];
  total: number;
  address: AddressInput;
}